+++ src/main/java/cc/oakk/lotto/view/printer/impl/LottosPrinter.java
@@ -1,7 +1,9 @@
-package cc.oakk.lotto.view.printer;
+package cc.oakk.lotto.view.printer.impl;
 
 import cc.oakk.lotto.model.Lotto;
 import cc.oakk.lotto.model.Lottos;
+import cc.oakk.lotto.view.printer.Printer;
+import cc.oakk.lotto.view.printer.StringPrinter;
 
 import java.util.stream.Collectors;
